added scaling from model to views

+++ app/src/main/java/org/example/event/PositionListener.java
@@ -1,5 +1,5 @@
 package org.example.event;
 
 public interface PositionListener {
-    void onPositionChanged(int x, int y);
+    void onPositionChanged(double x, double y);
 }
