+++ app/src/main/java/org/example/model/Model.java
@@ -1,6 +1,6 @@
 package org.example.model;
 
-public record Model(Integer width, Integer height, Position position, MeasuringTape measuringTape) {
+public record Model(Double width, Double height, Position position, MeasuringTape measuringTape) {
 
     public void moveTo(Position position){
         this.position.moveTo(position.x(), position.y());
