+++ app/src/main/java/org/example/model/Position.java
@@ -6,17 +6,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Position {
-    private Integer x, y;
-    private int maxX = Integer.MAX_VALUE;
-    private int maxY = Integer.MAX_VALUE;
+    private double x, y;
+    private double maxX = Double.MAX_VALUE;
+    private double maxY = Double.MAX_VALUE;
     private final List<PositionListener> listeners = new ArrayList<>();
 
-    public Position(Integer x, Integer y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setBounds(int maxX, int maxY) {
+    public void setBounds(double maxX, double maxY) {
         this.maxX = maxX;
         this.maxY = maxY;
     }
@@ -25,7 +25,7 @@ public class Position {
         listeners.add(listener);
     }
 
-    public void moveTo(Integer x, Integer y) {
+    public void moveTo(double x, double y) {
         this.x = Math.max(0, Math.min(x, maxX));
         this.y = Math.max(0, Math.min(y, maxY));
         for (PositionListener l : listeners) {
@@ -33,6 +33,6 @@ public class Position {
         }
     }
 
-    public Integer x() { return this.x; }
-    public Integer y() { return this.y; }
+    public double x() { return this.x; }
+    public double y() { return this.y; }
 }
